pyproject.toml
[project]
name = "synglot"
version = "0.1.0"
description = "Synthetic Data Generation and Translation Toolkit"
readme = "README.md"
requires-python = ">=3.8"
license = {text = "MIT"}
authors = [
    {name = "Ilija Lichkovski", email = "ilija@manifold.mk"},
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Text Processing :: Linguistic",
]

dependencies = [
    "transformers",
    "datasets",
    "torch",
    "PyYAML",
    "numpy",
    "sacremoses",
    # Add other dependencies as needed, e.g.:
    # "openai",
    # "anthropic",
    # "google-generativeai",
]

[project.urls]
Homepage = "https://github.com/yourusername/synglot" # Replace with your repo URL
Repository = "https://github.com/yourusername/synglot" # Replace with your repo URL
"Bug Tracker" = "https://github.com/yourusername/synglot/issues" # Replace with your repo URL

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["synglot*"] # Adjust if your package is not in the root or named synglot
exclude = ["docs*", "tests*", "outputs*"] 